+++ source/error.c
@@ -44,9 +44,9 @@ bool typecheck(LispObject * obj, LispObject * pred_sym) {
 
 // found_bug
 // Handle a bug.
-void found_bug(const char * file, int line, const char * func) {
+void found_bug(const char * file, long line, const char * func) {
     // TODO: instructions for reporting
-    printf("\n%s, line %d, in %s:\nIt looks like you have found a bug!\n\n",
+    printf("\n%s, line %ld, in %s:\nIt looks like you have found a bug!\n\n",
 	   file, line, func);
     exit(1);
 }
